restrict user actions

+++ QuizAppApi/QuizAppApi/Controllers/QuizController.cs
@@ -26,6 +26,10 @@ public class QuizController : ControllerBase
     [HttpPut("{id}")]
     public async Task<ActionResult<QuizManipulationResponseDto>> UpdateQuiz(int id, [FromBody] QuizManipulationRequestDto updateRequest)
     {
+        if (!(await _quizService.CanUserEditQuizAsync((User?)HttpContext.Items["User"], id)))
+        {
+            return Forbid();
+        }
         return await _quizService.UpdateQuizAsync(id, updateRequest);
     }
 
@@ -55,6 +59,10 @@ public class QuizController : ControllerBase
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteQuiz(int id)
     {
+        if (!(await _quizService.CanUserEditQuizAsync((User?)HttpContext.Items["User"], id)))
+        {
+            return Forbid();
+        }
         bool response = await _quizService.DeleteQuizAsync(id);
         if (response)
         {
